Apply patches without composer.json declaration
* Apply all patches contained in `composer-patches` directory.
* Output `patch` error message on fail.
* Remove shell patch support.
